$/Engineering/ProductDevelopment/AttributeFinder/AFUtils/Code $/Engineering/RC.Net/Utilities/ContextTags/Core/Code $/Engineering/ReusableComponents/BaseUtils/Code $/Engineering/ReusableComponents/COMComponents/UCLIDCOMUtils/Core/Code
https://extract.atlassian.net/browse/ISSUE-12851
Change the HANDLE parameter on EditCustomtags and EditTags to type int.  The HANDLE type doesn’t get marshalled properly for .NET

Steve made the changes and I reviewed them

William

[git-vault-id] Extract$/Engineering@32651/84804

+++ RC.Net/Utilities/Core/Code/PathTagsBase.cs
@@ -304,7 +304,7 @@ namespace Extract.Utilities
         /// <param name="hParentWindow">If not <see langword="null"/>, the tag editing UI will be
         /// displayed modally this window; otherwise the editor window will be modeless.</param>
         [SuppressMessage("Microsoft.Naming", "CA1725:ParameterNamesShouldMatchBaseDeclaration", MessageId = "0#")]
-        public virtual void EditCustomTags(IntPtr hParentWindow)
+        public virtual void EditCustomTags(int hParentWindow)
         {
             try
             {
